requirements.txt
tqdm
tabulate
numpy
torch>=0.3
scipy
Pillow
torchvision
cython
matplotlib
scikit-image
tensorflow-gpu==1.7.0
keras>=2.0.8
opencv-python
h5py
imgaug
IPython[all]
